+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0-alpha.2] - 2021-11-23
+
 ### Changed
 
 - `RawAddress` (`Address` at v0.1.x) struct has been renamed to `BdAddr`.
@@ -40,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release
 
-[Unreleased]: https://github.com/yskszk63/bdaddr/compare/v0.2.0-alpha.1...HEAD
+[Unreleased]: https://github.com/yskszk63/bdaddr/compare/v0.2.0-alpha.2...HEAD
+[0.2.0-alpha.1]: https://github.com/yskszk63/bdaddr/compare/v0.2.0-alpha.1...v0.2.0-alpha.2
 [0.2.0-alpha.1]: https://github.com/yskszk63/bdaddr/compare/v0.1.2...v0.2.0-alpha.1
 [0.1.2]: https://github.com/yskszk63/bdaddr/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/yskszk63/bdaddr/compare/v0.1.0...v0.1.1